Moved ReefTarget to its own file
