+++ src/main/java/frc/robot/Constants.java
@@ -40,23 +40,6 @@ public final class Constants {
     RIGHT
   }
 
-  public static enum ReefTarget {
-    TWO(9, 22),
-    FOUR(8, 17),
-    SIX(7, 18),
-    EIGHT(6, 19),
-    TEN(11, 20),
-    TWELVE(10, 21);
-
-    public final double red;
-    public final double blue;
-
-    ReefTarget(double red, double blue) {
-      this.red = red;
-      this.blue = blue;
-    }
-  }
-
   public static class OperatorConstants {
     public static final int kDriverControllerPort = 0;
     public static final String canivoreSerial = "D75CCE723353385320202034111303FF";
