+++ src/main/java/frc/robot/RobotContainer.java
@@ -27,7 +27,6 @@ import edu.wpi.first.wpilibj2.command.button.CommandGenericHID;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine;
 import frc.robot.Constants.Direction;
-import frc.robot.Constants.ReefTarget;
 import frc.robot.commands.DriveCommands;
 import frc.robot.commands.boathookCommands.BoathookExtendMotionPathCommand;
 import frc.robot.commands.boathookCommands.BoathookRetractMotionPathCommand;
@@ -37,6 +36,7 @@ import frc.robot.commands.elasticCommands.PreCheckTab;
 import frc.robot.commands.intakeCommands.*;
 import frc.robot.commands.visionCommands.GoToReefCommand;
 import frc.robot.commands.visionCommands.TargetChange;
+import frc.robot.game.ReefTarget;
 import frc.robot.generated.TunerConstants;
 import frc.robot.subsystems.RumbleSubsystem;
 import frc.robot.subsystems.boathook.Boathook;
@@ -85,18 +85,15 @@ public class RobotContainer {
   // each of these corresponds to a different button on the button board
   // these should set the pipeline to the side of the reef where the button is located
   // numbers correspond to clock faces with twelve being the back face of the reef
-  private final TargetChange twelveLeft =
-      new TargetChange(Direction.LEFT, 180, ReefTarget.TWELVE);
+  private final TargetChange twelveLeft = new TargetChange(Direction.LEFT, 180, ReefTarget.TWELVE);
   private final TargetChange twelveRight =
       new TargetChange(Direction.RIGHT, 180, ReefTarget.TWELVE);
 
   private final TargetChange tenLeft = new TargetChange(Direction.LEFT, -120, ReefTarget.TEN);
   private final TargetChange tenRight = new TargetChange(Direction.RIGHT, -120, ReefTarget.TEN);
 
-  private final TargetChange eightLeft =
-      new TargetChange(Direction.LEFT, -60, ReefTarget.EIGHT);
-  private final TargetChange eightRight =
-      new TargetChange(Direction.RIGHT, -60, ReefTarget.EIGHT);
+  private final TargetChange eightLeft = new TargetChange(Direction.LEFT, -60, ReefTarget.EIGHT);
+  private final TargetChange eightRight = new TargetChange(Direction.RIGHT, -60, ReefTarget.EIGHT);
 
   private final TargetChange sixLeft = new TargetChange(Direction.LEFT, 0, ReefTarget.SIX);
   private final TargetChange sixRight = new TargetChange(Direction.RIGHT, 0, ReefTarget.SIX);
